Fixed the fix of CID 578018

+++ src/ssl.c
@@ -40281,7 +40281,7 @@ void* wolfSSL_GetDhAgreeCtx(WOLFSSL* ssl)
                 WOLFSSL_MSG("Serial size error");
                 return WOLFSSL_FAILURE;
             }
-            if (sizeof(cert->serial) < serialSz) {
+            if ((int)sizeof(cert->serial) < serialSz) {
                 WOLFSSL_MSG("Serial buffer too small");
                 return BUFFER_E;
             }
